Strangler extraction #3: worker bodies -> zxnu_workers.py; classic-sync test; v9.0.8
Moves ~780 lines of pure worker-thread code out of MainWindow.__init__ into
zxnu_workers.py, next to its listen-mode sibling run_remote_listen_server:

- run_classic_sync_server: the classic Sync3/Sync4 server loop (was
  nextsync_do_server_job's inner loop). Every UI tentacle is now an injected
  callable - log, sync_once/always_sync/get_conflict_policy (0-arg, read per
  use so mid-session Settings flips still apply), ask_conflict (blocking
  overwrite prompt), set_session_active, pane_progress. MainWindow keeps a
  ~60-line adapter under the historical nextsync_do_server_job name.
- hdfmonkey transfer worker bodies: _run_get_task / _run_put_task /
  _run_put_external_task / _run_delete_task + the recursive scan helpers +
  is_directory, parameterized on execute (execute_hdf_monkey) and
  check_full_disk; all call sites updated.
- sync framing + helpers: sendpacket/recv_block/recv_exact/
  sanitize_incoming_path, getFileList/agecheck/update_syncpoint/timestamp.
- is_filetype_a_directory now has a single home in zxnu_config (the pane and
  the workers import it; the MainWindow closure is gone).

New tests/test_classic_sync.py: 19 localhost checks driving the real server
with a mock dot - Sync4 handshake, chunked PC->Next pull honouring
max_payload + packet numbering, Next->PC framed upload, overwrite/ignore
conflict policies, syncpoint bookkeeping and hostile-path sanitation. This
protocol had zero coverage before. Wired into run_all.py.

Fixes found in passing:
- The 'Use slow transfer' setting never worked: its MAX_PAYLOAD rebind was a
  function-local no-op. Payload size is now derived from the persisted
  setting at each server start, so the toggle actually switches to 256-byte
  payloads.

Also: generalized the star-import tripwire (asserts all three star-imported
modules carry the catch-all __all__; zxnu_api adopted it), and bumped
ZX_NEXT_UNITE_VERSION 9.0.7 -> 9.0.8.

Suite green (7 files, 7 UI phases); ruff clean.
